Packed external nullifier encoding

+++ world-chain-builder/crates/world/pool/src/builder.rs
@@ -22,14 +22,14 @@ use crate::validator::WorldChainTransactionValidator;
 /// config.
 #[derive(Debug, Clone)]
 pub struct WorldChainPoolBuilder {
-    pub num_pbh_txs: u16,
+    pub num_pbh_txs: u8,
     pub pbh_validator: Address,
     pub pbh_signature_aggregator: Address,
 }
 
 impl WorldChainPoolBuilder {
     pub fn new(
-        num_pbh_txs: u16,
+        num_pbh_txs: u8,
         pbh_validator: Address,
         pbh_signature_aggregator: Address,
     ) -> Self {
